+++ src/lib/components/chat/MessageInput.svelte
@@ -557,7 +557,6 @@
 									<InputMenu
 										bind:webSearchEnabled
 										bind:selectedToolIds
-										{availableToolIds}
 										uploadFilesHandler={() => {
 											filesInputElement.click();
 										}}
